% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaschSim.R
\name{RaschSim}
\alias{RaschSim}
\title{RaschSim() Function}
\usage{
RaschSim(items, persons)
}
\arguments{
\item{items:}{the number of items}

\item{persons:}{the number of persons}

\item{i.loc:}{generating item difficulty parameter}

\item{p.loc:}{generating person ability parameter}

\item{resp:}{generating personXitem matrix}
}
\value{

}
\description{
This function generates dataset which fits the Rasch measurement model
}
\examples{
   x<- RaschSim(10, 100)
   x
}
