% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaschSim.R
\name{RaschSim}
\alias{RaschSim}
\title{RaschSim() Function}
\usage{
RaschSim(items, persons)
}
\arguments{
\item{items:}{the number of items}

\item{persons:}{the number of persons}
}
\value{
i.loc: generating item difficulty parameter

p.loc: generating person ability parameter

resp: generating personXitem matrix
}
\description{
This function generates dataset which fits the Rasch measurement model
}
\examples{
   x<- RaschSim(10, 100)
   x
}
