% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaschSim.R
\name{RaschSim}
\alias{RaschSim}
\title{RaschSim() Function}
\usage{
RaschSim(items, persons)
}
\arguments{
\item{items, persons}{: two numbers to be operated}
}
\description{
This function generates dataset which fits the Rasch measurement model
}
\examples{
RaschSim(10, 100)
}
\keyword{RaschSim}
